Addition of Collision Geometry functionality
Added functionality to calculate Monte Carlo Glaube MC simulations of the collision geometry of nucleus-nucleus collisions. The position and type of nucleons is kept in objects of class Nucleon. The class Nucleus is a container of nucleons. The class CollisionGeometry is a container of two nuclei and the coordinates of nn interactions. The class CollisionGemetryHistogram produces histograms of the number of participants, number of binary collisions, and the eccentricity of the overlap zone. The classes CollisionGeometryGenerator and CollisionGeometryAnalyzer are inheriting from the class Task and define tasks that can be inserted in a usual event loops. The class CollisionGeometryConfiguration contains parameters used for simulations and analysis of collision geometry. There are different types of nuclear geometry that are possible -- Woods-Saxon,  Gaussian, hard sphere, etc. The app RunCollisionGeometrySimulator provides an example of usage  for the geomtry package.